backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
httpx==0.28.1
anthropic==0.43.0
openai>=1.0.0
PyJWT==2.12.1
psycopg[binary]==3.3.3
langchain>=0.3.20
langchain-core>=0.3.20
langchain-community>=0.3.20
faiss-cpu==1.10.0
sentence-transformers==3.4.1
slowapi==0.1.9
llm-guard==0.3.14
numpy<2
duckduckgo-search>=6.3.7
pytest==8.4.2
pytest-asyncio==0.26.0
rich==14.2.0
